requirements.txt
glfw==2.8.0
nodeenv==1.9.1
numpy==2.2.3
pybullet==3.2.7
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyright==1.1.396
scipy==1.15.2
typing_extensions==4.12.2
